refactor(test): extract chat host fixture
